ai-traffic-analysis/backend/requirements.txt
fastapi>=0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
opencv-python-headless==4.10.0.84
ultralytics==8.3.0
Pillow>=10.4.0
sqlalchemy==2.0.35
python-dotenv==1.0.1
numpy>=1.26.4
aiofiles==24.1.0
pydantic>=2.9.2
